+++ shared/src/main/scala/com/cibo/evilplot/plot/renderers/BarRenderer.scala
@@ -33,10 +33,11 @@ package com.cibo.evilplot.plot.renderers
 import com.cibo.evilplot.colors.Color
 import com.cibo.evilplot.geometry._
 import com.cibo.evilplot.plot.aesthetics.Theme
-import com.cibo.evilplot.plot.{Bar, Plot}
+import com.cibo.evilplot.plot.{Bar, LegendContext, Plot}
 
 trait BarRenderer extends PlotElementRenderer[Bar] {
   def render(plot: Plot, extent: Extent, category: Bar): Drawable
+  def legendContext: Option[LegendContext] = None
 }
 
 object BarRenderer {
@@ -47,6 +48,27 @@ object BarRenderer {
   )(implicit theme: Theme): BarRenderer = new BarRenderer {
     def render(plot: Plot, extent: Extent, bar: Bar): Drawable = {
       Rect(extent.width, extent.height).filled(color.getOrElse(theme.colors.bar))
+    }
+  }
+
+  /** A BarRenderer that assigns a single name to this bar. */
+  def named(
+    color: Option[Color] = None,
+    name: Option[String] = None,
+    legendElement: Option[Drawable] = None
+  )(implicit theme: Theme): BarRenderer = new BarRenderer {
+    def render(plot: Plot, extent: Extent, bar: Bar): Drawable = {
+      Rect(extent.width, extent.height).filled(color.getOrElse(theme.colors.bar))
+    }
+
+    override def legendContext: Option[LegendContext] = name.map { n =>
+      LegendContext.single(
+        element = legendElement.getOrElse {
+          val legSize = theme.fonts.legendLabelSize
+          Rect(legSize, legSize).filled(color.getOrElse(theme.colors.bar))
+        },
+        label = n
+      )
     }
   }